Fix subsequent load issue
There was an issue where load() call would not resolve if attempting to create
new instance when all other instances are destroyed. This was because we were
attempting to reload the script a second time which Youtube wont fire onYouTubeIframeAPIReady
more than once, so the load() call was never resolving.